Refactor to modular Omesh Super-Dashboard
