+++ omesh_dashboard/pages/__init__.py
@@ -1,0 +1,9 @@
+from . import sales, google_analytics, google_ads, social
+
+__all__ = [
+    "sales",
+    "google_analytics",
+    "google_ads",
+    "social",
+]
+
